// packages/contracts/generated/AnnotationRevision.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnnotationDocument } from "./AnnotationDocument";
import type { Id } from "./Id";

export type AnnotationRevision = { annotation_revision_id: Id, parent_revision_id: Id | null, revision_no: number, document: AnnotationDocument, created_at: string, created_by: Id, content_hash: string, };